Changed `delete_flag_handling()` logic in utils.bash, now it deletes only the lines present in benchIdX.txt (also deletes the lines from the txt file).
A LOT of refactoring duplicated code to `common/`.

Functional runme.sh for benchId02 using `common/autotune_find_best_comb.py`.

Added autotune.py for benchId02.

+++ benchmarks/common/autotune_find_best_comb.py
@@ -1,109 +1,42 @@
-from matplotlib import pyplot as plt
 import argparse
+import copy
+import pandas as pd
+import sys
+from colorama import init, Fore
 import json
 import pathlib
-import pandas as pd
-import seaborn as sns
-import sys
-from matplotlib.lines import Line2D
-from colorama import init, Fore
-import re
-
 
 sys.path.append(str(pathlib.Path(__file__).resolve().parents[2].joinpath('common')))
 from timer_stats import TimerStatsParser
+from utils import *
 
 
-def get_all_hw_names(dump_dir: str, bench_id: str) -> [str]:
-    """
-    Get the abs path of all the json files in the dump directory, recursively.
-    """
-    # check if f'benchId{bench_id}.txt' exists
-    if not pathlib.Path(dump_dir).joinpath(f'benchId{bench_id}.txt').exists():
-        print(f'File {dump_dir}/benchId{bench_id}.txt does not exist.')
-        return []
-    # read all the lines from the file at dump_dir/benchId.txt
-    with open(pathlib.Path(dump_dir).joinpath(f'benchId{bench_id}.txt'), 'r') as f:
-        lines = f.readlines()
-
-    hw_names = []
-    for line in lines:
-        # skip if line is empty
-        if not line.strip():
-            continue
-        # find all the json files in current_run_dir
-        line = line.strip()
-        # seperate hw, path for each line
-        hw, line = line.split(',', 1)
-        if hw not in hw_names:
-            hw_names.append(hw)
-
-    print(f'Found these hardware names: {hw_names}')
-    return hw_names
-
-
-def get_all_json_files(dump_dir: str, bench_id: str, only_this_hw: str) -> [str]:
-    """
-    Get the abs path of all the json files in the dump directory, recursively.
-    """
-    if not pathlib.Path(dump_dir).joinpath(f'benchId{bench_id}.txt').exists():
-        print(f'File {dump_dir}/benchId{bench_id}.txt does not exist.')
-        return []
-    # read all the lines from the file at dump_dir/benchId.txt
-    with open(pathlib.Path(dump_dir).joinpath(f'benchId{bench_id}.txt'), 'r') as f:
-        lines = f.readlines()
-
-    json_files = []
-    for line in lines:
-        # skip if line is empty
-        if not line.strip():
-            continue
-
-        # find all the json files in line
-        line = line.strip()
-
-        # seperate hw, path for each line
-        hw, sub_dump_dir_name = line.split(',', 1)
-        sub_dump_dir_name = sub_dump_dir_name.strip()
-
-        if only_this_hw != hw:
-            continue
-
-        sub_dump_dir_path = pathlib.Path(dump_dir) / pathlib.Path(sub_dump_dir_name)
-        json_files.extend([str(f) for f in sub_dump_dir_path.rglob('*.json')])
-
-    print(f'Found {len(lines)} sub-dump directories and a '
-          f'total of {len(json_files)} json files for benchmark ID {bench_id}')
-    return json_files
-
-
-def get_best_config(dumps_dir: str, benchid: str, out: str):
+def get_best_config(dumps_dir: str, benchid: str, out: str, parse_pairs_func=lambda pairs: {}, unique_names_two_args_func=lambda pairs, hw: ''):
     """
     Find the best unroll factor for each hardware and print it.
     It looks for benchId{benchid}.txt in the dumps_dir to extract all the relevant json files.
     """
     init()
     all_hw_names = get_all_hw_names(dumps_dir, benchid)
+
+    # json_report structure: (no lists, everything is a dict)
+    # benchId -> hw_name -> N -> config -> {key: value}
+    json_report_dict = {}
+
     for hw_name in all_hw_names:
         print('=================================================')
         print('Finding best unroll factor for hardware:', hw_name)
 
-        all_jsons = get_all_json_files(dumps_dir, benchid, hw_name)
+        unique_names_func = lambda x: unique_names_two_args_func(x, hw_name)
 
-        parse_pairs = lambda pairs: {
-            'N': pairs['N'],
-            'unroll_factor': pairs['unroll_factor']
-        }
-        unique_names = lambda pairs: \
-            'name=' + pairs['name'] + f';;hw={hw_name}' + ';;N=' + str(pairs['N']) + ';;unroll=' + str(
-                pairs['unroll_factor'])
+        all_jsons = get_all_json_files(dumps_dir, benchid, hw_name)
         parse_unique_names = lambda name: {
             pair.split('=')[0]: pair.split('=')[1] for pair in name.split(';;')
         }
 
-        parsed_runs = [TimerStatsParser(j, parse_pairs) for j in all_jsons]
+        parsed_runs = [TimerStatsParser(j, parse_pairs_func) for j in all_jsons]
         parsed_union = pd.concat([run.get_df() for run in parsed_runs], ignore_index=True)
-        parsed_union['name_N_hw_comb'] = parsed_union.apply(unique_names, axis=1)
+        parsed_union['name_N_hw_comb'] = parsed_union.apply(unique_names_func, axis=1)
 
         # phase 1: Extract the unique name_N_hw_comb s.
         unique_name_N_hw_comb = parsed_union['name_N_hw_comb'].unique()
@@ -120,7 +53,13 @@ def get_best_config(dumps_dir: str, benchid: str, out: str):
             if key not in configs_and_runtimes:
                 configs_and_runtimes[key] = []
 
-            filtered_conf = {'unroll': conf['unroll']}
+            filtered_conf = copy.copy(conf)
+
+            # Filter out hw, N, benchId and keep the rest
+            filtered_conf.pop('hw')
+            filtered_conf.pop('N')
+            filtered_conf.pop('name')
+
             configs_and_runtimes[key].append((filtered_conf, median_runtime))
 
         # phase 3: Find and print the best unroll factor and all configurations for each case
@@ -135,19 +74,79 @@ def get_best_config(dumps_dir: str, benchid: str, out: str):
                 for conf, runtime in sorted_configs
             ])
 
-            msg = f'Best configuration for {Fore.GREEN}benchId{benchid}{Fore.RESET}, kernel ({Fore.GREEN}{name}{Fore.RESET}), on {Fore.GREEN}{hw}{Fore.RESET}, for N={Fore.GREEN}{N}{Fore.RESET} is {Fore.RED}{best_config}{Fore.RESET} FROM {Fore.YELLOW}[{all_configs_str}]{Fore.RESET}'
+            msg = (f'Best configuration for {Fore.GREEN}benchId{benchid}{Fore.RESET}, '
+                   f'kernel ({Fore.GREEN}{name}{Fore.RESET}), on {Fore.GREEN}{hw}{Fore.RESET}, '
+                   f'for N={Fore.GREEN}{N}{Fore.RESET} is {Fore.RED}{best_config}{Fore.RESET} '
+                   f'FROM {Fore.YELLOW}[{all_configs_str}]{Fore.RESET}'
+            )
             print(msg)
 
-            with open(out, 'w') as f:
-                plain_text = re.sub(r'\033\[[0-9;]*m', '', msg)
-                f.write(plain_text + '\n')
+            # Create keys if they don't exist in json_report
+            if benchid not in json_report_dict:
+                json_report_dict[benchid] = {}
+            if hw not in json_report_dict[benchid]:
+                json_report_dict[benchid][hw] = {}
+            if str(N) not in json_report_dict[benchid][hw]:
+                json_report_dict[benchid][hw][str(N)] = {}
+            json_report_dict[benchid][hw][str(N)] = best_config
+
+    json_report = json.dumps(json_report_dict, indent=4)
+    # check if out file exists, if exists, read it and merge with json_report_dict and write it back (overwrite)
+    # if not exists, create it and write json_report_dict
+    # if there is conflict, throw exception
+
+    if pathlib.Path(out).exists():
+        print(f'Merging the new json report with the existing one at {out}')
+        with open(out, 'r') as f:
+            existing_json_report = json.load(f)
+
+        # merge the two dictionaries
+        for bench_id, hw_dict in json_report_dict.items():
+            if bench_id not in existing_json_report:
+                existing_json_report[bench_id] = {}
+
+            for hw, N_dict in hw_dict.items():
+                if hw not in existing_json_report[bench_id]:
+                    existing_json_report[bench_id][hw] = {}
+
+                for N, config in N_dict.items():
+                    if N not in existing_json_report[bench_id][hw]:
+                        existing_json_report[bench_id][hw][N] = config
+                    else:
+                        # Check if the existing configuration matches the new one
+                        existing_config = existing_json_report[bench_id][hw][N]
+                        if existing_config != config:
+                            raise Exception(
+                                f"Conflict in the existing JSON file {out} for benchId {bench_id}, hw {hw}, N {N}.\n"
+                                f"Existing: {existing_config}, New: {config}"
+                            )
+        # If no exception was raised, the dictionaries are merged successfully.
+
+        # write the merged dictionary back to the file
+        with open(out, 'w') as f:
+            json.dump(existing_json_report, f, indent=4)
+    else:
+        print(f'Writing the new json report to {out}')
+        with open(out, 'w') as f:
+            f.write(json_report)
 
 
 if __name__ == '__main__':
     # Can be used as a standalone script
     parser = argparse.ArgumentParser()
     parser.add_argument('--dumps-dir', type=str, required=True)
-    parser.add_argument('--out', type=str, required=True)
     parser.add_argument('--benchid', type=str, required=True)
     args = parser.parse_args()
-    get_best_config(args.dumps_dir, args.benchid, args.out)
+    get_best_config(
+        args.dumps_dir,
+        args.benchid,
+        (pathlib.Path(args.dumps_dir)/"autotuner.json").__str__(),
+        lambda pairs: {
+           'N': pairs['N'],
+           'UNROLL_FACTOR0': pairs['UNROLL_FACTOR0']
+        },
+        lambda pairs, hw_name: \
+           'name=' + pairs['name'] + f';;hw={hw_name}' + ';;N=' + str(pairs['N']) + ';;UNROLL_FACTOR0=' + str(
+               pairs['UNROLL_FACTOR0']
+        )
+    )
